Optimized accelerometer values

+++ autoino-android/src/de/kulawik/autoino/accelerometer/AccelerometerListener.java
@@ -1,5 +1,5 @@
 package de.kulawik.autoino.accelerometer;
 
 public interface AccelerometerListener {
-	public void onAccelerationChanged(float x, float y, float z);
+	public void onAccelerationChanged(int x, int y, int z);
 }
